feat: added userid to verify flashcards belongs to correct user

+++ src/main/java/com/flashcard/flashcard_api/service/FlashcardService.java
@@ -2,7 +2,9 @@ package com.flashcard.flashcard_api.service;
 
 import com.flashcard.flashcard_api.entity.Flashcard;
 import com.flashcard.flashcard_api.repository.FlashcardRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Date;
 import java.util.List;
@@ -18,6 +20,7 @@ public class FlashcardService {
 
     public Flashcard createFlashcard(Flashcard flashcard, String userId) {
         flashcard.setCreatedAt(new Date());
+        flashcard.setUserId(userId);
         return repository.save(flashcard);
     }
 
@@ -25,12 +28,19 @@ public class FlashcardService {
         return repository.findByUserId(userId);
     }
 
-    public Flashcard getFlashcardById(String id) {
-        return repository.findById(id).orElse(null);
+    public Flashcard getFlashcardById(String id, String userId) {
+        Flashcard flashcard = repository.findById(id).orElse(null);
+        if (!flashcard.getUserId().equals(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.FORBIDDEN,
+                    "This flashcard does not belong to you!"
+            );
+        }
+        return flashcard;
     }
 
-    public Flashcard updateFlashcard(String id, Flashcard updated) {
-        Flashcard existing = getFlashcardById(id);
+    public Flashcard updateFlashcard(String id, Flashcard updated, String userId) {
+        Flashcard existing = getFlashcardById(id, userId);
         if (existing != null) {
             existing.setQuestion(updated.getQuestion());
             existing.setAnswer(updated.getAnswer());
@@ -39,7 +49,14 @@ public class FlashcardService {
         return null;
     }
 
-    public void deleteFlashcard(String id) {
+    public void deleteFlashcard(String id, String userId) {
+        Flashcard flashcard = getFlashcardById(id, userId);
+        if (!flashcard.getUserId().equals(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.FORBIDDEN,
+                    "This flashcard does not belong to you!"
+            );
+        }
         repository.deleteById(id);
     }
 }
